Test: SellerInfo 관련 TC 분리
