+++ mission/tests/unit/ItemInfo/ItemInfo.spec.js
@@ -9,126 +9,6 @@ describe('ItemInfoPage', () => {
     expect(wrapper.find('#item-info-page').exists()).toBe(true);
   });
 
-  describe('Seller Section', () => {
-    const wrapper = mount(ItemInfoPage);
-
-    describe('Seller Name', () => {
-      it('renders seller-name', async () => {
-        expect(wrapper.find('[data-test="seller-name"]').exists()).toBeTruthy();
-      });
-
-      it('displays seller-name from data', async () => {
-        const sellerName = '테스터';
-        await wrapper.setData({
-          itemInfo: {
-            sellerName,
-          },
-        });
-
-        expect(wrapper.find('[data-test="seller-name"]').text()).toEqual(sellerName);
-      });
-    });
-
-    describe('Profile Img', () => {
-      const profileImg = 'testImg';
-      it('renders profile-img when there is a profile image', async () => {
-        await wrapper.setData({
-          itemInfo: {
-            profileImg,
-          },
-        });
-
-        expect(wrapper.find('[data-test="profile-img"]').exists()).toBeTruthy();
-      });
-
-      it('displays profile-img from data', async () => {
-        await wrapper.setData({
-          itemInfo: {
-            profileImg,
-          },
-        });
-
-        expect(wrapper.find('[data-test="profile-img"]').attributes().src).toEqual(profileImg);
-      });
-
-      it('renders default-profile-img when is there is no profile image', async () => {
-        await wrapper.setData({
-          itemInfo: {
-            profileImg: undefined,
-          },
-        });
-
-        expect(wrapper.find('[data-test="default-profile-img"]').exists()).toBeTruthy();
-      });
-    });
-
-    describe('Tag', () => {
-      it('renders tags when there is tags', async () => {
-        const tags = ['남성', '의류', '10대'];
-        await wrapper.setData({
-          itemInfo: {
-            tags,
-          },
-        });
-
-        expect(wrapper.findAll('[data-test="seller-tag"]').length).toEqual(tags.length);
-      });
-
-      it('hides tags when there is no tags', async () => {
-        const tags = [];
-        await wrapper.setData({
-          itemInfo: {
-            tags,
-          },
-        });
-
-        expect(wrapper.find('[data-test="seller-tag"]').exists()).toBeFalsy();
-      });
-
-      it('renders default-profile-img when is there is no profile image', async () => {
-        await wrapper.setData({
-          itemInfo: {
-            profileImg: undefined,
-          },
-        });
-
-        expect(wrapper.find('[data-test="default-profile-img"]').exists()).toBeTruthy();
-      });
-    });
-
-    describe('Favorite Button', () => {
-      beforeEach(async () => {
-        await wrapper.setData({
-          isFavorite: false,
-        });
-      });
-      it('renders favorite-btn', () => {
-        expect(wrapper.find('[data-test="favorite-btn"]').exists()).toBeTruthy();
-      });
-
-      it('is initially rendered as regular style', () => {
-        expect(wrapper.find('[data-test="favorite-btn"]').attributes().icon).toContain('far');
-      });
-
-      it('changes to solid style when clicked', async () => {
-        const button = wrapper.find('[data-test="favorite-btn"]');
-
-        await button.trigger('click');
-
-        expect(button.attributes().icon).toContain('fas');
-      });
-
-      it('returns to the original style by double clicking', async () => {
-        const button = wrapper.find('[data-test="favorite-btn"]');
-
-        await button.trigger('click');
-        await button.trigger('click');
-
-        expect(button.attributes().icon).toContain('far');
-      });
-    });
-  });
-
   describe('Price Section', () => {
     const wrapper = mount(ItemInfoPage);
     it('renders product-name', () => {
